+++ lib/src/templates/core_widgets_templates.dart
@@ -4218,16 +4218,16 @@ import 'package:{{project_name}}/exports.dart';
 class CustomSocialAuthButton extends StatelessWidget {
   const CustomSocialAuthButton({
     required this.text,
+    required this.iconPath,
     required this.onPressed,
     this.isLoading = false,
-    this.iconPath,
     super.key,
   });
 
   final String text;
+  final String iconPath;
   final VoidCallback? onPressed;
   final bool isLoading;
-  final String? iconPath;
 
   @override
   Widget build(BuildContext context) {
@@ -4235,13 +4235,11 @@ class CustomSocialAuthButton extends StatelessWidget {
       text: text,
       onPressed: isLoading ? null : onPressed,
       isLoading: isLoading,
-      prefixIcon: iconPath != null
-          ? SvgPicture.asset(
-              iconPath!,
-              width: 20,
-              height: 20,
-            )
-          : null,
+      prefixIcon: SvgPicture.asset(
+        iconPath,
+        width: 20,
+        height: 20,
+      ),
       iconSpacing: 12,
     );
   }